app/requirements.txt
# Base
six==1.15.0

# Django
django==3.0.5
djangorestframework==3.11.0

# DB
psycopg2-binary==2.8.5

# Testing
pytest-django==3.9.0
pytest==5.4.1
pytest-cov==2.8.1

# Server
gunicorn==20.0.4

# Envs
dj-database-url==0.5.0

# Static Files
whitenoise==5.0.1

# Code Quality
flake8===3.7.9  # https://flake8.pycqa.org/en/latest/
black==19.10b0  # https://black.readthedocs.io/en/stable
isort==4.3.21  # https://github.com/PyCQA/isort
